+++ czkawka_gui/src/main.rs
@@ -10,7 +10,6 @@ mod connect_button_save;
 mod connect_button_search;
 mod connect_button_select;
 mod connect_button_stop;
-mod connect_button_symlink;
 mod connect_compute_results;
 mod connect_header_buttons;
 mod connect_hide_text_view_errors;
@@ -52,7 +51,6 @@ use crate::connect_button_save::*;
 use crate::connect_button_search::*;
 use crate::connect_button_select::*;
 use crate::connect_button_stop::*;
-use crate::connect_button_symlink::*;
 use crate::connect_compute_results::*;
 use crate::connect_header_buttons::*;
 use crate::connect_hide_text_view_errors::*;
@@ -131,8 +129,7 @@ fn main() {
     );
     connect_button_select(&gui_data);
     connect_button_stop(&gui_data);
-    connect_button_symlink(&gui_data);
-    connect_button_hardlink(&gui_data);
+    connect_button_hardlink_symlink(&gui_data);
     connect_button_move(&gui_data);
     connect_notebook_tabs(&gui_data);
     connect_selection_of_directories(&gui_data);
